Added initializer code.
One of the problems with the initializer is that it has to use a
configuration setting in the app config file instead of allowing you to
pass in the connection string directly. Whereas the migrator allows you
to provide a connection string, or a connection string name (from the
config file).